+++ clientsrc/pages/user_index/mainarea/main.html
@@ -19,6 +19,7 @@
             :init-data="initData"
             :field-data="fieldData"
             :filed-title-map="saveField" 
+            :validator-options="validatorOptions"
             :is-add="isAdd"
             :title="saveTitle"
             :url="saveUrl">
